+++ index.js
@@ -1,5 +1,5 @@
 var next = require('./lib/index.js');
 
-next.version = '1.12.3';
+next.version = '1.13.0';
 
 module.exports = next;