+++ CHANGELOG.md
@@ -1,5 +1,81 @@
 # Change Log 
 
+<a name="1.13.0"></a>
+# [1.13.0](https://github.com/alibaba-fusion/next/compare/1.12.8...1.13.0) (2019-02-21)
+
+
+
+
+<a name="1.12.8"></a>
+## [1.12.8](https://github.com/alibaba-fusion/next/compare/1.12.7...1.12.8) (2019-02-21)
+
+
+
+
+<a name="1.12.7"></a>
+## [1.12.7](https://github.com/alibaba-fusion/next/compare/1.12.6...1.12.7) (2019-02-21)
+
+
+### Bug Fixes
+
+* **Radio:** child can contain null ([c703b60](https://github.com/alibaba-fusion/next/commit/c703b60))
+
+
+
+
+<a name="1.12.6"></a>
+## [1.12.6](https://github.com/alibaba-fusion/next/compare/1.12.5...1.12.6) (2019-02-21)
+
+
+### Bug Fixes
+
+* **Cascader:** set listStyle height style error ([61ea4f7](https://github.com/alibaba-fusion/next/commit/61ea4f7))
+* **Input:** add name to input dom ([51d151f](https://github.com/alibaba-fusion/next/commit/51d151f))
+
+
+
+
+<a name="1.12.5"></a>
+## [1.12.5](https://github.com/alibaba-fusion/next/compare/1.12.4...1.12.5) (2019-02-20)
+
+
+### Bug Fixes
+
+* style for rtl in Menu & Slider ([573694a](https://github.com/alibaba-fusion/next/commit/573694a))
+* **Message:** can`t config content font-size ([2c18b4d](https://github.com/alibaba-fusion/next/commit/2c18b4d))
+* **Radio:** a11y ([f26cb24](https://github.com/alibaba-fusion/next/commit/f26cb24))
+* **Table:** can't set className via getRowProps in tree mode, close [#343](https://github.com/alibaba-fusion/next/issues/343) ([44c20cb](https://github.com/alibaba-fusion/next/commit/44c20cb))
+
+
+### Features
+
+* **Tag:** wrap in ConfigProvider ([95233cc](https://github.com/alibaba-fusion/next/commit/95233cc))
+
+
+
+
+<a name="1.12.4"></a>
+## [1.12.4](https://github.com/alibaba-fusion/next/compare/1.12.3...1.12.4) (2019-02-14)
+
+
+### Bug Fixes
+
+* **Checkbox:** group didn't honor children's indeternimated state ([f90a747](https://github.com/alibaba-fusion/next/commit/f90a747))
+* **Input:** textarea placeholder in ie11 ([069c73b](https://github.com/alibaba-fusion/next/commit/069c73b))
+* **Select:** could add tag while popup not visible ([662a959](https://github.com/alibaba-fusion/next/commit/662a959))
+* **Table:** can't scroll horizontally when body scrollTop !== 0 ([5935b52](https://github.com/alibaba-fusion/next/commit/5935b52))
+* **Table:** head selection emptied wrongly ([adb3b59](https://github.com/alibaba-fusion/next/commit/adb3b59))
+
+
+### Features
+
+* **Icon:** suport size of inherit ([48e7f1d](https://github.com/alibaba-fusion/next/commit/48e7f1d))
+* **Progress:** remove tabindex ([602b267](https://github.com/alibaba-fusion/next/commit/602b267))
+* **Step:** uniform parameter of step ([e1d2ab5](https://github.com/alibaba-fusion/next/commit/e1d2ab5))
+
+
+
+
 <a name="1.12.3"></a>
 ## [1.12.3](https://github.com/alibaba-fusion/next/compare/1.12.2...1.12.3) (2019-01-28)